+++ benches/encrypted_message.rs
@@ -2,13 +2,13 @@ use criterion::{Criterion, black_box, criterion_group, criterion_main};
 use encrypted_message::{
     EncryptedMessage,
     strategy::{Deterministic, Randomized},
-    key_config::{KeyConfig, Secret},
+    config::{Config, Secret},
 };
 use rand::distributions::{Alphanumeric, DistString};
 
 #[derive(Debug, Default)]
-pub struct BenchesKeyConfig;
-impl KeyConfig for BenchesKeyConfig {
+pub struct BenchesConfig;
+impl Config for BenchesConfig {
     fn keys(&self) -> Vec<Secret<[u8; 32]>> {
         vec![(*b"uuOxfpWgRgIEo3dIrdo0hnHJHF1hntvW").into()]
     }
@@ -19,15 +19,15 @@ fn encrypted_message(c: &mut Criterion) {
     let payload = black_box(Alphanumeric.sample_string(&mut rand::thread_rng(), 32));
 
     c.bench_function("Encrypt 32-byte payload (Deterministic)", |b| b.iter(|| {
-        EncryptedMessage::<_, Deterministic, BenchesKeyConfig>::encrypt(payload.clone()).unwrap()
+        EncryptedMessage::<_, Deterministic, BenchesConfig>::encrypt(payload.clone()).unwrap()
     }));
 
     c.bench_function("Encrypt 32-byte payload (Randomized)", |b| b.iter(|| {
-        EncryptedMessage::<_, Randomized, BenchesKeyConfig>::encrypt(payload.clone()).unwrap()
+        EncryptedMessage::<_, Randomized, BenchesConfig>::encrypt(payload.clone()).unwrap()
     }));
 
     c.bench_function("Decrypt 32-byte payload", |b| {
-        let encrypted = EncryptedMessage::<_, Deterministic, BenchesKeyConfig>::encrypt(payload.clone()).unwrap();
+        let encrypted = EncryptedMessage::<_, Deterministic, BenchesConfig>::encrypt(payload.clone()).unwrap();
         b.iter(|| encrypted.decrypt().unwrap())
     });
 }