+++ examples/basic.rs
@@ -4,13 +4,13 @@
 use encrypted_message::{
     EncryptedMessage,
     strategy::Randomized,
-    key_config::{KeyConfig, Secret, ExposeSecret as _},
+    config::{Config, Secret, ExposeSecret as _},
 };
 
 /// NOTE: Never hardcode your keys like this, obviously.
 #[derive(Debug, Default)]
-struct AppKeyConfig;
-impl KeyConfig for AppKeyConfig {
+struct EncryptionConfig;
+impl Config for EncryptionConfig {
     fn keys(&self) -> Vec<Secret<[u8; 32]>> {
         let encoded_keys = [Secret::new("75754f7866705767526749456f33644972646f30686e484a484631686e747657".to_string())];
         encoded_keys.iter()
@@ -26,7 +26,7 @@ impl KeyConfig for AppKeyConfig {
 
 fn main() {
     // Encrypt a user's diary.
-    let diary: EncryptedMessage::<String, Randomized, AppKeyConfig> = {
+    let diary: EncryptedMessage::<String, Randomized, EncryptionConfig> = {
         EncryptedMessage::encrypt("Very personal stuff".to_string()).unwrap()
     };
     println!("Encrypted diary: {diary:#?}");