+++ src/testing.rs
@@ -1,8 +1,8 @@
-use crate::key_config::{KeyConfig, Secret};
+use crate::config::{Config, Secret};
 
 #[derive(Debug, Default, PartialEq, Eq)]
-pub struct TestKeyConfig;
-impl KeyConfig for TestKeyConfig {
+pub struct TestConfig;
+impl Config for TestConfig {
     fn keys(&self) -> Vec<Secret<[u8; 32]>> {
         vec![
             (*b"uuOxfpWgRgIEo3dIrdo0hnHJHF1hntvW").into(),