+++ tests/config_with_external_data.rs
@@ -1,19 +1,18 @@
 use encrypted_message::{
     EncryptedMessage,
     strategy::Randomized,
-    key_config::{KeyConfig, Secret},
+    config::{Config, Secret, ExposeSecret as _},
 };
 use pbkdf2::pbkdf2_hmac_array;
-use secrecy::{ExposeSecret as _, SecretString};
 use sha2::Sha256;
 
 #[derive(Debug)]
-struct UserKeyConfig {
-    user_password: SecretString,
-    salt: SecretString,
+struct UserEncryptionConfig {
+    user_password: Secret<String>,
+    salt: Secret<String>,
 }
 
-impl KeyConfig for UserKeyConfig {
+impl Config for UserEncryptionConfig {
     fn keys(&self) -> Vec<Secret<[u8; 32]>> {
         let raw_key = self.user_password.expose_secret().as_bytes();
         let salt = self.salt.expose_secret().as_bytes();
@@ -24,18 +23,18 @@ impl KeyConfig for UserKeyConfig {
 }
 
 #[test]
-fn key_config_with_external_data() {
-    let key_config = UserKeyConfig {
+fn config_with_external_data() {
+    let config = UserEncryptionConfig {
         user_password: "human-password-that-should-be-derived".to_string().into(),
         salt: "unique-salt".to_string().into(),
     };
 
     // Encrypt a payload.
-    let encrypted: EncryptedMessage<String, Randomized, UserKeyConfig> = {
-        EncryptedMessage::encrypt_with_key_config("Hi".to_string(), &key_config).unwrap()
+    let encrypted: EncryptedMessage<String, Randomized, UserEncryptionConfig> = {
+        EncryptedMessage::encrypt_with_config("Hi".to_string(), &config).unwrap()
     };
 
     // Decrypt the payload.
-    let decrypted = encrypted.decrypt_with_key_config(&key_config).unwrap();
+    let decrypted = encrypted.decrypt_with_config(&config).unwrap();
     assert_eq!(decrypted, "Hi");
 }